Suppresswarning added to as.numeric check in tfce.mixed, just like they were in vertex_analysis()
Additional lines were added to requirement checks, to make sure the script stops if the session is non-interactive and VWRfirstrun() flagged missing files. I cannot add a stop() inside the function as it will be interpreted as an error, so the functions that call VWRfirstrun() need to return something as well if files are missing.

VWRfirstrun() now creates an arbitrary string that gets created and caught by the upper functions in that situation. When created, this object is returned by the upper functions instead of continuing, practically stopping the function without error.

I also needed the ". <- NULL" line as CRAN doesn't like the manipulation of variables that don't exist "yet" in the environment