Custom path Python installation
VWRfirstrun() now implements a way for the user to customise where they want Python to be installed.

It calls python_custominstall(), a whole new script based on reticulate's python installation pipeline.

Packages cannot however be installed in that independent Python installation (reticulate imposes either conda or a virtual environment). Trying to make it install packages in the Python directory will return an error saying that the environment "exists but is not a virtual environment".

By default, brainstat or numpy will therefore be installed in that virtual environment. VWRfirstrun() was modified to check whether they exist in 1) the default environment 2) the virtual environment.

I have tried Examples 1 and 2 and they work when only a custom Python installation has been done.